+++ src/main/java/com/baro/jwt_auth/config/SecurityConfig.java
@@ -1,7 +1,7 @@
 package com.baro.jwt_auth.config;
 
-import com.baro.jwt_auth.common.response.ErrorCode;
-import com.baro.jwt_auth.common.response.ResErrorDTO;
+import com.baro.jwt_auth.common.exception.ErrorCode;
+import com.baro.jwt_auth.common.exception.dto.ResErrorDTO;
 import com.baro.jwt_auth.security.jwt.JwtAuthFilter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
